mlops-experiments/experiments/requirements.txt
mlflow>=2.9.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
boto3>=1.28.0
prometheus-client>=0.19.0
